Курсор последовательного порождения простых чисел
Close #1243
